feat: add review module, rename ProfileResponse to UserResponse, add reviews to vehicle detail

+++ internal/domain/entity/review.go
@@ -7,7 +7,7 @@ type Review struct {
 	UserID    int       `json:"user_id" gorm:"not null"`
 	VehicleID int       `json:"vehicle_id" gorm:"not null"`
 	RentalID  int       `json:"rental_id" gorm:"not null"`
-	Rating    int16     `json:"rating" gorm:"not null"`
+	Rating    int       `json:"rating" gorm:"not null"`
 	Comment   *string   `json:"comment"`
 	CreatedAt time.Time `json:"created_at"`
 	UpdatedAt time.Time `json:"updated_at"`
